# api/openapi.yaml
openapi: 3.0.3
info:
  title: HDF5 Agent API
  description: |
    Versioned HTTP API for browsing and manipulating HDF5 files.

    Other services in a data platform must call this API (JSON over HTTP)
    rather than linking against the HDF5 C library. The Go client in
    `pkg/hdf5client` is a convenience wrapper around these endpoints.

    All responses may include `X-Request-ID`. Clients should send that header
    to correlate logs across services. Errors use the envelope
    `{ "error": { "code", "message", "request_id" } }`.
  version: 1.1.0
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: /
    description: Current host
...
paths:
...
  /api/v1/files:
    get:
      tags: [files]
      operationId: listFiles
      summary: List HDF5 files in the data directory
      responses:
        "200":
          description: File inventory
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileList"
        "500":
          $ref: "#/components/responses/Error"
    post:
      tags: [files]
      operationId: createFile
      summary: Create an empty HDF5 file
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFile"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: created
                  name:
                    type: string
        "400":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
...
components:
...
  schemas:
...
    FileList:
      type: object
      required: [files]
      properties:
        files:
          type: array
          items:
            $ref: "#/components/schemas/FileInfo"
    FileInfo:
      type: object
      required: [name, size_bytes]
      properties:
        name:
          type: string
        size_bytes:
          type: integer
          format: int64
...
    CreateFile:
      type: object
      required: [name]
      properties:
        name:
          type: string
          example: experiment.h5
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum:
                - invalid_request
                - invalid_name
                - invalid_path
                - not_found
                - conflict
                - unsupported_type
                - too_large
                - not_ready
                - internal
            message:
              type: string
            request_id:
              type: string
  responses:
    Error:
      description: Structured error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"